feat: 유저 정보 조회에 마일리지 포함

+++ src/main/java/com/tg/coreservice/domain/Mileage.java
@@ -13,9 +13,9 @@ public class Mileage {
     @Id
     private Long user_id;
 
-    private Long mileage;
+    private int mileage;
 
-    public Mileage(Long post_id, Long mileage) {
+    public Mileage(Long post_id, int mileage) {
         this.user_id = post_id;
         this.mileage = mileage;
     }
